change them to ivec

+++ DragonicTactics/source/Game/DragonicTactics/StateComponents/GridSystem.cpp
@@ -57,26 +57,26 @@ bool GridSystem::IsOccupied(Math::ivec2 pos) const
 	return character_grid[pos.y][pos.x] != nullptr;
 }
 //A* algorithm
-int GridSystem::ManhattanDistance(Math::vec2 a, Math::vec2 b) const
+int GridSystem::ManhattanDistance(Math::ivec2 a, Math::ivec2 b) const
 {
 	return static_cast<int>(std::abs(a.x - b.x) + std::abs(a.y - b.y));
 }
 
-std::vector<Math::vec2> GridSystem::GetNeighbors(Math::vec2 position) const
+std::vector<Math::ivec2> GridSystem::GetNeighbors(Math::ivec2 position) const
 {
-	std::vector<Math::vec2> neighbors;
+	std::vector<Math::ivec2> neighbors;
 
 	// 4-directional movement(up,down,left,right)
-	const Math::vec2 directionals[] = {
+	const Math::ivec2 directionals[] = {
 		{  0,	 1 }, //  up
 		{  0,   -1 }, //  down
 		{ -1,    0 }, //  left
 		{  1,	 0 }  //  right
 	};
 
-	for (const Math::vec2& dir : directionals)
+	for (const Math::ivec2& dir : directionals)
 	{
-		Math::vec2 neighbor = position + dir;
+		Math::ivec2 neighbor = position + dir;
 		if (IsValidTile(neighbor))
 		{
 			neighbors.push_back(neighbor);
@@ -159,7 +159,7 @@ Character* GridSystem::GetCharacterAt(Math::ivec2 pos) const
 	return character_grid[pos.y][pos.x];
 }
 
-std::vector<Math::vec2> GridSystem::FindPath(Math::vec2 start, Math::vec2 goal)
+std::vector<Math::ivec2> GridSystem::FindPath(Math::ivec2 start, Math::ivec2 goal)
 {
 	// edge cases
 	if (!IsValidTile(start) || !IsValidTile(goal))
@@ -214,8 +214,8 @@ std::vector<Math::vec2> GridSystem::FindPath(Math::vec2 start, Math::vec2 goal)
 		}
 
 		//check neighbors
-		std::vector<Math::vec2> neighbors = GetNeighbors(current->position);
-		for (const Math::vec2& neighborPos : neighbors)
+		std::vector<Math::ivec2> neighbors = GetNeighbors(current->position);
+		for (const Math::ivec2& neighborPos : neighbors)
 		{
 			//skip if not walkable or in closed set
 			if (!IsWalkable(neighborPos))
@@ -254,21 +254,21 @@ std::vector<Math::vec2> GridSystem::FindPath(Math::vec2 start, Math::vec2 goal)
 	}
 
 	//reconstruct path
-	std::vector<Math::vec2> path;
+	std::vector<Math::ivec2> path;
 	
 
 
-	return std::vector<Math::vec2>();
+	return std::vector<Math::ivec2>();
 }
 
-int GridSystem::GetPathLength([[maybe_unused]] Math::vec2 start, [[maybe_unused]] Math::vec2 goal)
+int GridSystem::GetPathLength([[maybe_unused]] Math::ivec2 start, [[maybe_unused]] Math::ivec2 goal)
 {
 	return 0;
 }
 
-std::vector<Math::vec2> GridSystem::GetReachableTiles([[maybe_unused]] Math::vec2 start, [[maybe_unused]] int maxDistance)
+std::vector<Math::ivec2> GridSystem::GetReachableTiles([[maybe_unused]] Math::ivec2 start, [[maybe_unused]] int maxDistance)
 {
-	return std::vector<Math::vec2>();
+	return std::vector<Math::ivec2>();
 }
 
 void GridSystem::MoveCharacter(Math::ivec2 old_pos, Math::ivec2 new_pos)
